requirements.txt
appnope==0.1.2
APScheduler==3.6.3
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.7
debugpy==1.5.1
decorator==5.1.0
entrypoints==0.3
idna==3.3
ipykernel==6.5.0
ipython==7.29.0
jedi==0.18.0
jupyter-client==7.0.6
jupyter-core==4.9.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.1
numpy==1.21.4
pandas==1.3.4
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.22
ptyprocess==0.7.0
Pygments==2.10.0
python-dateutil==2.8.2
python-telegram-bot==13.8.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.3.0
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
tornado==6.1
traitlets==5.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
wcwidth==0.2.5
